refactor(ui): extract complex logic from productivity.tsx to useProductivityData hook
Extracted over 150 lines of complex filtering, sorting, derivation, and planning logic into a dedicated React hook `useProductivityData`. This drastically reduces the size of the `ProductivityScreen` component, decoupling domain logic from UI rendering and enforcing the Single Responsibility Principle. All unit tests, type checks, and linters passed successfully.
